Refactor: auth 화면을 공용 AuthHeader·(auth) 레이아웃으로 통일
- login·sign-up·forget-password·reset-password를 (auth) 라우트 그룹으로 모으고, 공용 layout이 <main> 랜드마크와 가운데 정렬 래퍼를 제공 (URL 불변)
- 공용 AuthHeader(뒤로가기 + 가운데 제목)로 세 화면 헤더 통일하고 중복 topbar 스타일 제거
- 로그인 leaf 로고 제거하고 큰 제목을 헤더 바로 이관
- reset-password를 로그인 모듈 의존에서 분리(자체 page.module.scss), 폴더 이동으로 깨진 PasswordUpdateForm의 actions import 경로 정정
