+++ src/app/(auth)/forget-password/_component/forgetPassword.module.scss
@@ -1,50 +1,6 @@
 // 비밀번호 찾기 2단계 흐름 (이메일 입력 → 인증 코드 입력).
 // UI만 — 코드 입력은 동작하는 목업(5분 타이머·재전송). 1단계 발송은 기존 액션 재사용.
-
-.wrap {
-  margin: $spacing-24 auto;
-  max-width: 43rem;
-  display: flex;
-  flex-direction: column;
-}
-
-// ── 상단 헤더 (뒤로가기 + 가운데 제목) ──
-
-.topbar {
-  display: flex;
-  align-items: center;
-  gap: $spacing-8;
-  margin-bottom: $spacing-16;
-}
-
-.back_button {
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  width: $spacing-40;
-  height: $spacing-40;
-  border: none;
-  border-radius: $radius-s;
-  background: none;
-  cursor: pointer;
-  font-size: $font-size-24;
-  color: $txt-primary;
-
-  @include hover-bg-shift($bg-hover);
-}
-
-.title {
-  flex: 1;
-  text-align: center;
-  font-size: $font-size-18;
-  font-weight: $font-weight-bold;
-  color: $txt-primary;
-}
-
-.topbar_spacer {
-  width: $spacing-40;
-  flex: 0 0 auto;
-}
+// 바깥 래퍼·가운데 정렬은 (auth)/layout이 소유한다.
 
 // ── 스텝 본문 ──
