+++ src/app/(auth)/forget-password/page.tsx
@@ -1,0 +1,5 @@
+import ForgetPasswordFlow from './_component/ForgetPasswordFlow';
+
+export default function ForgetPasswordPage() {
+  return <ForgetPasswordFlow />;
+}
