+++ src/app/(auth)/layout.module.scss
@@ -1,0 +1,5 @@
+// auth 공용 래퍼 — 가운데 정렬 콘텐츠 컬럼.
+.wrap {
+  margin: $spacing-40 auto;
+  max-width: 43rem;
+}
