+++ src/app/(auth)/login/page.module.scss
@@ -1,0 +1,57 @@
+.subtitle {
+  margin-bottom: $spacing-24;
+  font-size: $font-size-14;
+  line-height: $line-height-normal;
+  color: $txt-secondary;
+}
+
+.divide {
+  margin: $spacing-8 0;
+  position: relative;
+  text-align: center;
+
+  .divide_line {
+    position: absolute;
+    top: 50%;
+    left: 0;
+    width: 100%;
+    height: 1px;
+    background-color: $border-primary;
+  }
+
+  .caption {
+    position: relative;
+    padding: 0 $spacing-12;
+    background-color: $bg-primary;
+    font-size: $font-size-13;
+    color: $txt-tertiary;
+  }
+}
+
+.link_group {
+  margin-top: $spacing-20;
+  text-align: center;
+
+  a {
+    font-size: $font-size-14;
+    font-weight: $font-weight-semibold;
+    color: $txt-link;
+
+    @include hover-color-shift($primary-hover);
+  }
+}
+
+.signup {
+  margin-top: $spacing-40;
+  text-align: center;
+  font-size: $font-size-14;
+  color: $txt-tertiary;
+
+  a {
+    margin-left: $spacing-4;
+    font-weight: $font-weight-semibold;
+    color: $txt-link;
+
+    @include hover-color-shift($primary-hover);
+  }
+}
