+++ src/app/(auth)/reset-password/page.tsx
@@ -1,0 +1,14 @@
+import PasswordUpdateForm from '@/app/_component/auth/PasswordUpdateForm';
+import styles from './page.module.scss';
+
+export default function page() {
+  return (
+    <>
+      <div className={styles.header}>
+        <h1>비밀번호 재설정하기</h1>
+        <p>새로운 비밀번호를 입력하세요.</p>
+      </div>
+      <PasswordUpdateForm />
+    </>
+  );
+}
