+++ src/app/(auth)/sign-up/_component/CompleteStep.tsx
