+++ src/app/(auth)/sign-up/_component/InfoStep.tsx
