+++ src/app/(auth)/sign-up/_component/TermsStep.tsx
