+++ src/app/_component/auth/authHeader.module.scss
@@ -1,0 +1,37 @@
+// auth 화면 공용 상단 헤더 — 뒤로가기 + 가운데 제목 + 우측 spacer.
+
+.header {
+  display: flex;
+  align-items: center;
+  gap: $spacing-8;
+  margin-bottom: $spacing-16;
+}
+
+.back_button {
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  width: $spacing-40;
+  height: $spacing-40;
+  border: none;
+  border-radius: $radius-s;
+  background: none;
+  cursor: pointer;
+  font-size: $font-size-24;
+  color: $txt-primary;
+
+  @include hover-bg-shift($bg-hover);
+}
+
+.title {
+  flex: 1;
+  text-align: center;
+  font-size: $font-size-18;
+  font-weight: $font-weight-bold;
+  color: $txt-primary;
+}
+
+.spacer {
+  width: $spacing-40;
+  flex: 0 0 auto;
+}
