--- src/app/forget-password/page.tsx
@@ -1,12 +1,0 @@
-import { LayoutContainer } from '@/components/layout';
-import ForgetPasswordFlow from './_component/ForgetPasswordFlow';
-
-export default function ForgetPasswordPage() {
-  return (
-    <section>
-      <LayoutContainer>
-        <ForgetPasswordFlow />
-      </LayoutContainer>
-    </section>
-  );
-}
