--- src/app/login/page.module.scss
@@ -1,112 +1,0 @@
-.wrap {
-  margin: $spacing-40 auto;
-  max-width: 40rem;
-}
-
-.topbar {
-  margin-bottom: $spacing-16;
-}
-
-.back_button {
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  width: $spacing-40;
-  height: $spacing-40;
-  margin-left: calc(-1 * $spacing-8);
-  border: none;
-  border-radius: $radius-s;
-  background: none;
-  cursor: pointer;
-  font-size: $font-size-24;
-  color: $txt-primary;
-
-  @include hover-bg-shift($bg-hover);
-}
-
-.logo {
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  width: $spacing-48;
-  height: $spacing-48;
-  margin-bottom: $spacing-16;
-  border-radius: $radius-s;
-  background-color: $accent;
-  color: $txt-inverse;
-
-  svg {
-    width: $spacing-24;
-    height: $spacing-24;
-  }
-}
-
-.header {
-  margin-bottom: $spacing-24;
-
-  h1 {
-    margin-bottom: $spacing-8;
-    font-size: $font-size-26;
-
-    @include respond-up($breakpoint-tablet-sm) {
-      font-size: $font-size-34;
-    }
-  }
-
-  p {
-    font-size: $font-size-14;
-    line-height: $line-height-normal;
-    color: $txt-secondary;
-  }
-}
-
-.divide {
-  margin: $spacing-8 0;
-  position: relative;
-  text-align: center;
-
-  .divide_line {
-    position: absolute;
-    top: 50%;
-    left: 0;
-    width: 100%;
-    height: 1px;
-    background-color: $border-primary;
-  }
-
-  .caption {
-    position: relative;
-    padding: 0 $spacing-12;
-    background-color: $bg-primary;
-    font-size: $font-size-13;
-    color: $txt-tertiary;
-  }
-}
-
-.link_group {
-  margin-top: $spacing-20;
-  text-align: center;
-
-  a {
-    font-size: $font-size-14;
-    font-weight: $font-weight-semibold;
-    color: $txt-link;
-
-    @include hover-color-shift($primary-hover);
-  }
-}
-
-.signup {
-  margin-top: $spacing-40;
-  text-align: center;
-  font-size: $font-size-14;
-  color: $txt-tertiary;
-
-  a {
-    margin-left: $spacing-4;
-    font-weight: $font-weight-semibold;
-    color: $txt-link;
-
-    @include hover-color-shift($primary-hover);
-  }
-}
